additional-studies/mini-project/ai-dev-deployment/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.7.1+cu128
torchvision==0.22.1+cu128
torchaudio==2.7.1+cu128
pillow==11.0.0
bentoml==1.4.15
fastapi==0.115.12
uvicorn==0.34.3
transformers==4.52.4
prometheus_fastapi_instrumentator==7.1.0
GPUtil==1.4.0
paramiko==3.5.1
scp==0.15.0
pyyaml==6.0.2
mlflow==2.22.0